add abstraction and fix add zero to output
