+++ src/main/java/edu/unc/ceccr/chembench/persistence/Descriptors.java
@@ -13,6 +13,14 @@ public class Descriptors implements java.io.Serializable {
     public Descriptors() {
     }
 
+    public void addDescriptorValues(List<Double> descriptorValues){
+        this.descriptorValues.addAll(descriptorValues);
+    }
+
+    public Double getDescriptorValueAtIndex(int index){
+        return descriptorValues.get(index);
+    }
+
     public List<Double> getDescriptorValues() {
         return descriptorValues;
     }
